requirements.txt
python-docx
docxcompose
pyinstaller